+++ instrumentation/opentelemetry-instrumentation-celery/src/opentelemetry/instrumentation/celery/utils.py
@@ -61,8 +61,11 @@ def set_attributes_from_context(span, context):
 
         # Skip `timelimit` if it is not set (it's default/unset value is a
         # tuple or a list of `None` values
-        if key == "timelimit" and value in [(None, None), [None, None]]:
-            continue
+        if key == "timelimit":
+            if value in [(None, None), [None, None]]:
+                continue
+            if None in value:
+                value = ["" if tl is None else tl for tl in value]
 
         # Skip `retries` if it's value is `0`
         if key == "retries" and value == 0: